@@ -1,94 +1,55 @@
-var React = require('react');
-var Events = require('../../lib/Events.js');
-var classNames = require('classnames');
+import React from 'react';
 
-var TransformButtons = [
+const Events = require('../../lib/Events.js');
+const classNames = require('classnames');
+const transformButtons = [
   { value: 'translate', icon: 'fa-arrows-alt' },
   { value: 'rotate', icon: 'fa-repeat' },
   { value: 'scale', icon: 'fa-expand' }
 ];
 
-export default class TransformToolbar extends React.Component {
-  constructor (props) {
-    super(props);
-    this.state = {
-      selectedTransform: 'translate',
-      localSpace: false
-    };
-  }
+export default () => {
+  const [ selectedTransform, setSelectedTransform ] = React.useState('translate');
+  const [ localSpace, setLocalSpace ] = React.useState(false);
 
-  componentDidMount () {
+  React.useEffect(() => {
     Events.on('transformmodechange', mode => {
-      this.setState({ selectedTransform: mode });
+      setSelectedTransform(mode);
     });
 
     Events.on('transformspacechange', () => {
-      Events.emit(
-        'transformspacechanged',
-        this.state.localSpace ? 'world' : 'local'
-      );
-      this.setState({ localSpace: !this.state.localSpace });
+      Events.emit('transformspacechanged', localSpace ? 'world' : 'local');
+      setLocalSpace(!localSpace);
     });
-  }
+  }, []);
 
-  changeTransformMode = mode => {
-    this.setState({ selectedTransform: mode });
+  const changeTransformMode = mode => {
+    setSelectedTransform(mode);
     Events.emit('transformmodechange', mode);
-    ga('send', 'event', 'Toolbar', 'selectHelper', mode);
   };
 
-  onLocalChange = e => {
-    const local = e.target.checked;
-    this.setState({ localSpace: local });
-    Events.emit('transformspacechanged', local ? 'local' : 'world');
-  };
-
-  renderTransformButtons = () => {
-    return TransformButtons.map(
-      function (option, i) {
-        var selected = option.value === this.state.selectedTransform;
-        var classes = classNames({
+  const renderTransformButtons = () => {
+    return transformButtons.map((option, i) => {
+        const classes = classNames({
           button: true,
           fa: true,
           [option.icon]: true,
-          active: selected
+          active: option.value === selectedTransform
         });
 
         return (
           <a
             title={option.value}
             key={i}
-            onClick={this.changeTransformMode.bind(this, option.value)}
+            onClick={() => changeTransformMode(option.value)}
             className={classes}
           />
         );
-      }.bind(this)
+      }
     );
   };
 
-  render () {
-    return (
-      <div id="transformToolbar" className="toolbarButtons">
-        {this.renderTransformButtons()}
-        {/* <span className="local-transform">
-          <input
-            id="local"
-            type="checkbox"
-            title="Toggle between local and world space transforms"
-            checked={
-              this.state.localSpace || this.state.selectedTransform === 'scale'
-            }
-            disabled={this.state.selectedTransform === 'scale'}
-            onChange={this.onLocalChange}
-          /> */}
-          {/* <label
-            htmlFor="local"
-            title="Toggle between local and world space transforms"
-          >
-            local
-          </label> */}
-        {/* </span> */}
-      </div>
-    );
-  }
-}
+  return <div id="transformToolbar" className="toolbarButtons">
+    {renderTransformButtons()}
+  </div>;
+};
