@@ -537,6 +537,8 @@ export function printEntity (entity, onDoubleClick) {
     return '';
   }
 
+debugger;
+
   // Icons.
   let icons = '';
   for (let objType in ICONS) {
@@ -557,24 +559,12 @@ export function printEntity (entity, onDoubleClick) {
     type = 'mixin';
   }
 
+  console.log('--------------------------');
+
   return (
-    <span className="entityPrint" onDoubleClick={onDoubleClick}>
-      <span className="entityTagName">
-        {'<' + entity.tagName.toLowerCase()}
-      </span>
-      {entityName && (
-        <span className="entityName" data-entity-name-type={type}>
-          &nbsp;{entityName}
-        </span>
-      )}
-      {!!icons && (
-        <span
-          className="entityIcons"
-          dangerouslySetInnerHTML={{ __html: icons }}
-        />
-      )}
-      <span className="entityCloseTag">{'>'}</span>
-    </span>
+
+    <span>{entityName}</span>
+
   );
 }
 
